requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Dance==7.1.0
Werkzeug==3.1.3
python-dotenv==1.1.1
requests==2.32.5
Pillow==11.3.0
email-validator==2.3.0
authlib==1.6.3
Flask-APScheduler==1.13.1
WTForms==3.2.1
stripe==12.5.1
itsdangerous==2.2.0
email-validator==2.3.0
openai==1.107.0
langchain==0.3.27
langchain-openai==0.3.32
langchain-community==0.3.29
psycopg2-binary==2.9.10
gunicorn==23.0.0